+++ src/builders/dev-server/dev-server-utils.ts
@@ -20,31 +20,23 @@ export function applySkyuxDevServerOptions(options: SkyuxDevServerBuilderOptions
   options.sslCert = getCertPath('skyux-server.crt');
   options.sslKey = getCertPath('skyux-server.key');
 
-  // Set options specific to SKY UX Host.
-  if (options.skyuxLaunch === 'host') {
-    const localUrl = getLocalUrlFromOptions(options);
+  const localUrl = getLocalUrlFromOptions(options);
 
-    // Point live-reloading back to localhost.
-    options.publicHost = localUrl;
-    options.allowedHosts = ['.blackbaud.com'];
+  // Point live-reloading back to localhost.
+  options.publicHost = localUrl;
+  options.allowedHosts = ['.blackbaud.com'];
 
-    // Point lazy-loaded modules to the localhost URL.
-    options.deployUrl = localUrl;
-    options.servePath = '/';
+  // Point lazy-loaded modules to the localhost URL.
+  options.deployUrl = localUrl;
+  options.servePath = '/';
 
-    // Disable Angular CLI's opening behavior since the Host URL Webpack plugin
-    // handles launching the browser.
-    options.open = false;
+  // Disable Angular CLI's opening behavior since the Host URL Webpack plugin
+  // handles launching the browser.
+  options.open = false;
 
-    // Default to opening the SKY UX Host URL.
-    /* istanbul ignore else */
-    if (options.skyuxOpen === undefined) {
-      options.skyuxOpen = true;
-    }
-  }
-
-  // Open the user's default browser automatically.
-  if (options.skyuxLaunch === 'local') {
-    options.open = true;
+  // Default to opening the SKY UX Host URL.
+  /* istanbul ignore else */
+  if (options.skyuxOpen === undefined) {
+    options.skyuxOpen = true;
   }
 }