+++ src/tools/webpack/host-asset.ts
@@ -1,10 +1,24 @@
+import {
+  SkyuxHostAssetType
+} from './host-asset-type';
+
 export interface SkyuxHostAsset {
 
   /**
-   * The name of the global "fallback" variable to test if the asset file has loaded.
-   * @see: https://docs.microsoft.com/en-us/aspnet/core/mvc/views/tag-helpers/built-in/link-tag-helper
+   * The name of the global "fallback" variable to test if the JavaScript file has loaded.
+   * @see: https://docs.microsoft.com/en-us/aspnet/core/mvc/views/tag-helpers/built-in/script-tag-helper
    */
   fallback?: string;
+
+  /**
+   * The CSS fallback rule to test if the style sheet file has loaded.
+   * @see: https://docs.microsoft.com/en-us/aspnet/core/mvc/views/tag-helpers/built-in/link-tag-helper
+   */
+  fallbackStylesheet?: {
+    class: string;
+    property: string;
+    value: string;
+  };
 
   /**
    * Specifies if the chunk is the main chunk for the entry point.
@@ -16,5 +30,10 @@ export interface SkyuxHostAsset {
    * The name of the file.
    */
   name: string;
+
+  /**
+   * The type of asset.
+   */
+  type: SkyuxHostAssetType;
 
 }