+++ src/tools/webpack/plugins/open-host-url/open-host-url.plugin.ts
@@ -54,7 +54,8 @@ export class SkyuxOpenHostUrlPlugin {
           host: this.config.host,
           localUrl: this.config.localUrl,
           rootElementTagName: 'app-root',
-          scripts: assets
+          scripts: assets.scripts,
+          stylesheets: assets.stylesheets
         };
 
         if (this.config.externals) {
@@ -63,7 +64,7 @@ export class SkyuxOpenHostUrlPlugin {
 
         const url = createHostUrl(
           this.config.host.url,
-          this.config.pathName,
+          this.config.baseHref,
           hostUrlConfig
         );